@@ -7,16 +7,24 @@ from process_bigraph import Composite
 
 from biosimulator_processes.steps.viz import CompositionPlotter, Plotter2d
 from biosimulator_processes.data_model.sed_data_model import MODEL_TYPE
-from biosimulator_processes.utils import register_module
+from biosimulator_processes.helpers import register_module
 
 
 # Define a list of processes to attempt to import and register
 PROCESSES_TO_REGISTER = [
-    ('cobra', 'cobra_process.CobraProcess'),
-    ('copasi', 'copasi_process.CopasiProcess'),
-    # ('smoldyn', 'smoldyn_process.SmoldynProcess'),
-    ('tellurium', 'tellurium_process.TelluriumProcess'),
-    ('amici', 'amici_process.AmiciProcess')]
+    ('cobra-process', 'cobra_process.CobraProcess'),
+    ('copasi-process', 'copasi_process.CopasiProcess'),
+    ('tellurium-process', 'tellurium_process.TelluriumProcess'),
+    ('utc-amici', 'amici_process.UtcAmici'),
+    ('utc-copasi', 'copasi_process.UtcCopasi')]
+
+
+try:
+    import smoldyn
+    PROCESSES_TO_REGISTER.append(('smoldyn-process', 'smoldyn_process.SmoldynProcess'))
+except:
+    print('Smoldyn is not properly installed in this environment and thus its process implementation cannot be registered. Please consult smoldyn documentation.')
+
 
 STEPS_TO_REGISTER = [
     ('copasi-step', 'ode_simulation.CopasiStep'),
@@ -31,5 +39,5 @@ CORE = ProcessTypes()
 
 # core type system implementation (unique to this package)
 CORE.type_registry.register('sed_model', schema={'_type': MODEL_TYPE})
-register_module(PROCESSES_TO_REGISTER, CORE)
-register_module(STEPS_TO_REGISTER, CORE)
+register_module(PROCESSES_TO_REGISTER, CORE, verbose=False)
+register_module(STEPS_TO_REGISTER, CORE, verbose=True)
